A working project. We can select and schedule task.
There are a lot of changes in this commit.
Most of the code is a boilerplate, though. Essentially,
we're either queueing a task by using its delay method
or adding a new PeriodicTask entry to djcelery's crontab.

That is, the most important file to look is views.py,
and it's also worth to do a tiny peek at the ScheduleField
class in forms.py. The rest isn't really Celery-related code.